Move PipelineRunner to toplevel sdk package (automated refactor)
This allows excluding the runner-author-only sdk.runners package from the
public javadoc.
